Add a new CLI script to dump given conversation
This is typically used for putting the chat log into our General
developer meeting pages. Previously, Michael used to do this manually
via copy/paste and processing in Excel (which, as he admitted, was quite
time consuming job). The current version supports MediaWiki output only
for now, with the possibility to extend in the future if needed. The
output still needs manual checking (we may want to surround whole
message text into <nowiki> in the future probably).