Add GHA support
Adding 401, 402 and master by default (php80 and php81).
Only with MySQL (that should be enough).

Note that the phpdoc tests are disabled because they have
a good number of missing documentation.

Also, phpcs emits a bunch of warnings that should be fixed
in order to enable --max-warnings 0

https://github.com/moodlehq/moodle-local_chatlogs/issues/13 has
been created about that.